refactor: <MenuIcon> to <SidebarItem>
and update project's README to reflect the new changes
