+++ client/src/components/slots/SidebarItem.vue
